refactor(responsemanager): reorganize for clarity
Clarify actor pattern and responsibilities across threads
